use jwt to check auth

+++ routers/post.go
@@ -24,7 +24,7 @@ func GetPosts(c *gin.Context) {
 		200,
 		gin.H{
 			"code": code,
-			"msg": "", //todo
+			"msg": myError.GetMsg(code),
 			"data": ret,
 		})
 
@@ -49,17 +49,26 @@ func GetPost(c *gin.Context) {
 //新建post
 func AddPost(c *gin.Context) {
 	var post models.Post
-	err := c.BindJSON(&post) // json的解析
 	statusCode := 200
 	code := myError.SUCCESS
-	if err != nil {
-		code = myError.INVALID_PARAMS
-		statusCode = 400
-	}
-	err = models.CreatePost(post)
-	if err != nil {
-		code = myError.INVALID_PARAMS
-		statusCode = 400
+	user, ok  := c.Get("user")
+	if !ok {
+		code = myError.CONTEXT_GET_USER_ERROR
+
+	} else {
+		err := c.BindJSON(&post) // json的解析
+		if err != nil {
+			code = myError.INVALID_PARAMS
+			statusCode = 400
+		}
+		u, _ := user.(models.User) // 接口的类型断言，这里将接口类型转换成了User struct
+		post.UserId = u.ID
+		err = models.CreatePost(post)
+		if err != nil {
+			code = myError.INVALID_PARAMS
+			statusCode = 400
+		}
+
 	}
 
 	c.JSON(statusCode, gin.H{
@@ -71,18 +80,34 @@ func AddPost(c *gin.Context) {
 func EditPost(c *gin.Context) {
 	var post models.Post
 	err := c.BindJSON(&post)
+
 	id := com.StrTo(c.Param("id")).MustInt()
 	statusCode := 200
 	code := myError.SUCCESS
+
 
 	if err != nil {
 		code = myError.INVALID_PARAMS
 		statusCode = 400
 	}
-	models.EditPost(id, post)
+	user, _  := c.Get("user")
+	u, _ := user.(models.User)
+	p := models.GetPost(id)
+	// 检查是否修改自己的文章
+	if u.ID != p.UserId {
+		code = myError.PERMISSION_DENIED
+		statusCode = 403
+	} else {
+		update := make(map[string]interface{})
+		//只有title和description两个字段允许修改
+		update["title"] = post.Title
+		update["description"] = post.Description
+		models.EditPost(id, update)
+	}
+
 	c.JSON(statusCode, gin.H{
 		"code": code,
-		"msg": myError.GetMsg(code), //todo
+		"msg": myError.GetMsg(code),
 	})
 
 }
@@ -92,12 +117,22 @@ func DeletePost(c *gin.Context) {
 	code := myError.SUCCESS
 	id := com.StrTo(c.Param("id")).MustInt()
 	isExist := models.IsPostExist(id)
+	statusCode := 200
 	if !isExist {
 		code = myError.PostNotFound
 	} else {
-		models.DeletePost(id)
+		user, _  := c.Get("user")
+		u, _ := user.(models.User)
+		p := models.GetPost(id)
+		// 检查操作的文章是否属于该用户
+		if u.ID != p.UserId {
+			code = myError.PERMISSION_DENIED
+			statusCode = 403
+		} else {
+			models.DeletePost(id)
+		}
 	}
-	c.JSON(200, gin.H{
+	c.JSON(statusCode, gin.H{
 		"code": code,
 		"msg": myError.GetMsg(code),
 	})
